arreglo de errores, terminado el trabajo

+++ Controller/MarksController.php
@@ -1,28 +1,21 @@
 <?php
 
-    require_once "./View/ProductsView.php";
     require_once "./View/MarksView.php";
-    require_once "./Model/ProductsModel.php";
     require_once "./Model/MarksModel.php";
     require_once "./View/LoginView.php";
-    require_once "./Controller/LoginController.php";
+    require_once "./View/ProductsView.php";
+    require_once "Helper.php";
 
-    class MarksController{
+    class MarksController extends Helper{
 
-        private $productsView;
         private $marksView;
-        private $productosModel;
         private $marksModel;
-        private $loginView;
-        private $loginControl;
+        private $productsView;
 
         function __construct(){
-            $this->productsView = new ProductsView();
             $this->marksView = new MarksView();
-            $this->productosModel = new ProductsModel();
             $this->marksModel = new MarksModel();
-            $this->loginView = new LoginView();
-            $this->loginControl = new LoginController();
+            $this->productsView = new ProductsView();
         }
         //LLAMA AL HOME DE MARCAS
         function HomeMarks(){
@@ -31,8 +24,8 @@
         }
         //INSERTA UNA NUEVA MARCA
         function InsertMark(){
-            $logeado = $this->loginControl->checkLoggedIn();
-            if($logeado){
+            $logeado = $this->CheckLoggedIn();
+            if($logeado && $_SESSION['ADMIN'] == 1){
                 if (isset($_POST['input_mark']) && isset($_POST['input_category'])) {
                     $mark = $_POST['input_mark'];
                     $category = $_POST['input_category'];
@@ -40,35 +33,35 @@
                 }
                 $this->productsView->ShowLocation('admin');
             }else{
-                $this->loginView->ShowLogin();
+                $this->productsView->ShowLocation('login');
             }
         }
         //ELIMINA UNA MARCA POR ID
         function DeleteMark($params = null){
-            $logeado = $this->loginControl->checkLoggedIn();
-            if($logeado){
+            $logeado = $this->CheckLoggedIn();
+            if($logeado && $_SESSION['ADMIN'] == 1){
                 $mark_id = $params[':ID'];
                 $this->marksModel->DeleteMark($mark_id);
                 $this->productsView->ShowLocation('admin');
             }else{
-                $this->loginView->ShowLogin();
+                $this->productsView->ShowLocation('login');
             }
         }
         //LLAMA LA VISTA PARA EDITAR UNA MARCA POR ID
         function EditMark($params = null){
-            $logeado = $this->loginControl->checkLoggedIn();
-            if($logeado){
+            $logeado = $this->CheckLoggedIn();
+            if($logeado && $_SESSION['ADMIN'] == 1){
                 $mark_id = $params[':ID'];
                 $mark = $this->marksModel->GetMarkById($mark_id);
                 $this->marksView->ShowEditMark($mark);
             }else{
-                $this->loginView->ShowLogin();
+                $this->productsView->ShowLocation('login');
             }
         }
         //LLAMA A ACTUALIZAR UNA MARCA
         function UpdateMark($params = null){
-            $logeado = $this->loginControl->checkLoggedIn();
-            if($logeado){
+            $logeado = $this->CheckLoggedIn();
+            if($logeado && $_SESSION['ADMIN'] == 1){
                 $mark_id = $params[':ID'];
                 if (isset($_POST['edit_mark']) && isset($_POST['edit_category'])) {
                     $mark = $_POST['edit_mark'];
@@ -77,7 +70,7 @@
                 }
                 $this->productsView->ShowLocation('admin');
             }else{
-                $this->loginView->ShowLogin();
+                $this->productsView->ShowLocation('login');
             }
         }
     }
